Ensure default recipients for activities
One of those omnibus commits you feel bad about making.

Core change is adding a method to Activity to ensure that an activity
has at least _some_ recipients. If unspecified, sets recipients to
original post's recipients, or to followers if this is the first post.

Also, moved the check for recipients to its own method in Activity.