Scrubber class to scrub test, activities, and objects
Add a Scrubber utility to scrub user-generated HTML for XSS attacks.

Because there are lots of places in activities and object for HTML to
hide, and each of the data types has optional sub-objects that could
be hiding more HTML, the Scrubber class has scrubActivity() and
scrubObject() methods to make sure no rock goes unturned.

Includes a unit test, of course.