Add a group inbox feed endpoint
Add an inbox feed endpoint for the group inbox. This required tweaking
the activityFeed() function in lib/streams to take a context argument
rather than a user, which in turn caused a lot of downstream twiddling
with the user activity endpoints.

Also, create an inbox stream when the group is created.