Work around npm not building Jade for Git installs
After install, we check to see if there are as many built files as there
are templates. If there aren't, we run a build.

This will never run on normal installs, because the built files end up
in package tarballs. However, Git installs don't work like that, so we
have to manually run the build ourselves.